+++ example/windows/flutter/generated_plugin_registrant.cc
@@ -8,6 +8,7 @@
 
 #include <app_links/app_links_plugin_c_api.h>
 #include <audioplayers_windows/audioplayers_windows_plugin.h>
+#include <firebase_core/firebase_core_plugin_c_api.h>
 #include <flutter_onnxruntime/flutter_onnxruntime_plugin.h>
 #include <passkeys_windows/passkeys_windows_plugin.h>
 #include <pdfx/pdfx_plugin.h>
@@ -19,6 +20,8 @@ void RegisterPlugins(flutter::PluginRegistry* registry) {
       registry->GetRegistrarForPlugin("AppLinksPluginCApi"));
   AudioplayersWindowsPluginRegisterWithRegistrar(
       registry->GetRegistrarForPlugin("AudioplayersWindowsPlugin"));
+  FirebaseCorePluginCApiRegisterWithRegistrar(
+      registry->GetRegistrarForPlugin("FirebaseCorePluginCApi"));
   FlutterOnnxruntimePluginRegisterWithRegistrar(
       registry->GetRegistrarForPlugin("FlutterOnnxruntimePlugin"));
   PasskeysWindowsPluginRegisterWithRegistrar(